refactory kafka pubsub code to extract common kafka code for reuse
