+++ internal/component/kafka/kafka.go
@@ -1,0 +1,161 @@
+/*
+Copyright 2021 The Dapr Authors
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package kafka
+
+import (
+	"context"
+	"time"
+
+	"github.com/Shopify/sarama"
+	"github.com/dapr/kit/logger"
+	"github.com/dapr/kit/retry"
+)
+
+// Kafka allows reading/writing to a Kafka consumer group.
+type Kafka struct {
+	producer      sarama.SyncProducer
+	consumerGroup string
+	brokers       []string
+	logger        logger.Logger
+	authType      string
+	saslUsername  string
+	saslPassword  string
+	initialOffset int64
+	cg            sarama.ConsumerGroup
+	topics        map[string]bool
+	cancel        context.CancelFunc
+	consumer      consumer
+	config        *sarama.Config
+
+	backOffConfig        retry.Config
+	consumeRetryInterval time.Duration
+}
+
+func NewKafka(l logger.Logger) *Kafka {
+	return &Kafka{
+		logger: l,
+	}
+}
+
+// Init does metadata parsing and connection establishment.
+func (k *Kafka) Init(metadata map[string]string) error {
+	upgradedMetadata, err := k.upgradeMetadata(metadata)
+	if err != nil {
+		return err
+	}
+
+	meta, err := k.getKafkaMetadata(upgradedMetadata)
+	if err != nil {
+		return err
+	}
+
+	k.brokers = meta.Brokers
+	k.consumerGroup = meta.ConsumerGroup
+	k.initialOffset = meta.InitialOffset
+	k.authType = meta.AuthType
+
+	config := sarama.NewConfig()
+	config.Version = meta.Version
+	config.Consumer.Offsets.Initial = k.initialOffset
+
+	if meta.ClientID != "" {
+		config.ClientID = meta.ClientID
+	}
+
+	err = updateTLSConfig(config, meta)
+	if err != nil {
+		return err
+	}
+
+	switch k.authType {
+	case oidcAuthType:
+		k.logger.Info("Configuring SASL OAuth2/OIDC authentication")
+		err = updateOidcAuthInfo(config, meta)
+		if err != nil {
+			return err
+		}
+	case passwordAuthType:
+		k.logger.Info("Configuring SASL Password authentication")
+		k.saslUsername = meta.SaslUsername
+		k.saslPassword = meta.SaslPassword
+		updatePasswordAuthInfo(config, k.saslUsername, k.saslPassword)
+	case mtlsAuthType:
+		k.logger.Info("Configuring mTLS authentcation")
+		err = updateMTLSAuthInfo(config, meta)
+		if err != nil {
+			return err
+		}
+	}
+
+	k.config = config
+	sarama.Logger = SaramaLogBridge{daprLogger: k.logger}
+
+	k.producer, err = getSyncProducer(*k.config, k.brokers, meta.MaxMessageBytes)
+	if err != nil {
+		return err
+	}
+
+	k.topics = make(map[string]bool)
+
+	// Default retry configuration is used if no
+	// backOff properties are set.
+	if err := retry.DecodeConfigWithPrefix(
+		&k.backOffConfig,
+		metadata,
+		"backOff"); err != nil {
+		return err
+	}
+	k.consumeRetryInterval = meta.ConsumeRetryInterval
+
+	k.logger.Debug("Kafka message bus initialization complete")
+
+	return nil
+}
+
+func (k *Kafka) Close() (err error) {
+	k.closeSubscriptionResources()
+
+	if k.producer != nil {
+		err = k.producer.Close()
+		k.producer = nil
+	}
+
+	return err
+}
+
+func (k *Kafka) addTopic(newTopic string) []string {
+	// Add topic to our map of topics
+	k.topics[newTopic] = true
+
+	topics := make([]string, len(k.topics))
+
+	i := 0
+	for topic := range k.topics {
+		topics[i] = topic
+		i++
+	}
+
+	return topics
+}
+
+// EventHandler is the handler used to handle the subscribed event.
+type EventHandler func(ctx context.Context, msg *NewEvent) error
+
+// NewEvent is an event arriving from a message bus instance.
+type NewEvent struct {
+	Data        []byte            `json:"data"`
+	Topic       string            `json:"topic"`
+	Metadata    map[string]string `json:"metadata"`
+	ContentType *string           `json:"contentType,omitempty"`
+}
